backend/requirements.txt
django
djangorestframework
pyyaml
requests
psycopg2
Pillow
django-phone-field